refactor(#2456): split text_editor.py into models/validation/coordinator
- Extract 5 data classes to _text_editor_models.py (86 LOC)
- Extract validation methods to _URDFValidationMixin in
  _text_editor_validation.py (397 LOC)
- Slim text_editor.py coordinator to 655 LOC (was 1072 LOC);
  URDFTextEditor inherits from both mixins
- All public names re-exported for backward compatibility
- TDD contract tests added (11 tests, all GREEN)
